+++ src/containers/About/About.test.js
@@ -1,11 +1,11 @@
 import React from "react";
-import About from "./AboutContainer";
+import About from "./About";
 import {
   shallow
 } from "enzyme";
 
 //Need at least 1 test to pass Travis CI
-describe("<AboutContainer />", () => {
+describe("<About />", () => {
 
   it('creates component without crashing', () => {
     const about = < About / >
